clear models
