+++ open-hackathon-server/src/manager.py
@@ -5,7 +5,7 @@ This file is covered by the LICENSING file in the root of this project.
 from flask_script import Manager, Server, Shell
 
 from hackathon import app
-from hackathon.hmongo import drop_db
+from hackathon.hmongo.database import drop_db, setup_db, add_super_user
 from hackathon.hmongo.models import User
 
 banner = r"""
@@ -36,30 +36,14 @@ manager.add_command("shell", Shell(banner=banner, make_context=make_shell_contex
 @manager.command
 def reset_db():
     drop_db()
-    setup_db()
+    init_db()
 
 
 @manager.command
-def setup_db():
-    """Initialize db tables
-
-    make sure database and user correctly created in db
-    in case upgrade the table structure, the origin table need be dropped firstly
-    """
-    # init REQUIRED db data.
-
-    # reserved user is deleted, may not need in mongodb implementation
-
-    # default super admin
-
-    admin = User(
-        name="admin",
-        nickname="admin",
-        password="e8104164dfc4a479e42a9f6c0aefd2be",
-        avatar_url="/static/pic/monkey-32-32px.png",
-        is_super=True)
-
-    User.objects(name="admin").update_one(__raw__={"$set": admin.to_mongo().to_dict()}, upsert=True)
+def init_db():
+    setup_db()
+    # todo custom username and password
+    add_super_user("admin", "admin", "admin")
 
 
 if __name__ == "__main__":
